requirements.txt
agno
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
google-generativeai